Modify so that almost everything passes LCLint with the options in lclint.sh. This is actually worthwhile; I found and fixed a few bugs/edge cases while doing this. For more information on LCLint, see <http://lclint.cs.virginia.edu/>.
svn path=/trunk/yasm/; revision=335

+++ libyasm/preproc.h
@@ -43,7 +43,7 @@ struct preproc {
 
     /* Gets more preprocessed source code (up to max_size bytes) into buf.
      * Note that more than a single line may be returned in buf. */
-    int (*input) (char *buf, int max_size);
+    size_t (*input) (char *buf, size_t max_size);
 };
 
 /* Available preprocessors */
